requirements.txt
python-telegram-bot
requests
peewee
validators